requirements.txt
Django==4.0
pyTelegramBotAPI==4.11.0
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
psycopg2==2.8.6
djangorestframework==3.12.4
django-mptt==0.14.0

telebot~=0.0.5
googlemaps~=4.10.0